Spawn the extension host directly from the main process because fork() is now fast again
